ARM: dts: lpc18xx: add creg-clk node
Add node for the clock register found in the CREG block.
This allows us to control the low power 1 kHz and 32 kHz
clocks on LPC18xx.

Also replace the 32 kHz input clock on the CGU node with
the CREG 32 kHz clock. As this is how the clock is really
connected on LPC18xx.

Signed-off-by: Joachim Eastwood <manabian@gmail.com>