ASoC: atmel_ssc_dai: shift data on rising for DSP_A
If the CODEC samples data on falling edge the SSC must shift out
data on rising clock edge.

From datasheet:
 * Receive CKI:
   The data inputs (Data and Frame Sync signals) are sampled on
   Receive Clock falling edge.
 * Trasmit CKI:
   The data outputs (Data and Frame Sync signals) are shifted
   out on Transmit Clock rising edge. The Frame sync signal
   input is sampled on Transmit clock falling edge.

For a CODEC that shifts data out on rising edge and shifts
data in on falling edge the SSC must be configured with
different CKI values in the receiver and trasmitter.

Signed-off-by: Joachim Eastwood <manabian@gmail.com>